To build a model a translator is needed. This object associated with a model indicates how to translate and load model to or from a special language.
All classes are no more abstract. This way, there is no need to extend them into other project. The only need is to provide a translator.

Add 3 methods to get list of arc, node and attributes

A constructor was removed (File). This model must not take care of opening and reading a file. To build a model from a file, you must use ths construtor asking for a vector of string.