Define a new object : ColoaneProgress.
This object is used when the user call a service which needs some feedback. ColoaneProgress provides a progressbar monitor and a support for cancellation (which is not implemented yet)

Motor has been updated to use this new object.

Some synchronization have been implemented to take into account asynchronous calls and answers.